Add a new flag to disable setting query timeouts.
Unfortunately the Hive JDBC drivers are really bad, and trying to set
a query timeout on them will cause the driver to throw an exception.
HikariCP does this when it's trying to validate a DB connection prior
to returning it to the user. Though this is generally a good practice,
it means you can't use HikariCP with Hive. (Both HiveServer1 and
HiveServer2 have this issue.)